DiscreteData: update to new module structure
