+++ modules/discretedata/src/connectivity/connectioniterator.cpp
