+++ modules/discretedata/src/connectivity/euclideanmeasure.cpp
@@ -1,0 +1,60 @@
+/*********************************************************************************
+ *
+ * Inviwo - Interactive Visualization Workshop
+ *
+ * Copyright (c) 2012-2018 Inviwo Foundation
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ *
+ * 1. Redistributions of source code must retain the above copyright notice, this
+ * list of conditions and the following disclaimer.
+ * 2. Redistributions in binary form must reproduce the above copyright notice,
+ * this list of conditions and the following disclaimer in the documentation
+ * and/or other materials provided with the distribution.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+ * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+ * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
+ * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+ * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
+ * ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+ * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ *
+ *********************************************************************************/
+
+#include <modules/discretedata/connectivity/euclideanmeasure.h>
+
+namespace inviwo {
+namespace discretedata {
+namespace euclidean {
+
+double getMeasure(const Connectivity& grid, const Channel& positions, GridPrimitive dim,
+                  ind index) {
+
+    CellType cell = grid.getCellType(dim, index);
+    switch (cell) {
+        case CellType::Hexahedron:
+            if (positions.getNumComponents() == 3) {
+                // Only implemented 3D bodies so far.
+                if (dim != GridPrimitive::Volume) return -1;
+
+                double measure = -1;
+
+                measure = inviwo::dispatching::dispatch<double, dispatching::filter::Scalars>(
+                    positions.getDataFormatId(), HexVolumeComputer(), grid, positions, index);
+
+                return measure;
+            }
+        default:
+            return -1.0;
+    }
+}
+
+}
+}
+}
